0.4.4 Bugfix
Sorry I've been gone, I was in Ottawa on a tour

In This Update...
- Added a comment at Slate saying I need more time. It actually is pretty good, but when I copy-paste it it makes indents EVERYWHERE and ruins the code
- Fixed a bug where if you typed anything other than the commands coded, it would say HELP MODE. Type...
The reason this bug was there is because... I don't know. Whenever I do an or statement it does that. Weird.

Don't forget to file issues if you want something — a bugfix, feature... etc etc etc.